requirements.txt
# Python dependencies for face_recognition backend
face_recognition==1.3.0
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
dlib==19.24.2